--- monitor/README.md
@@ -1,1 +1,0 @@
-Placeholder for monitoring scripts
