+++ sapmon/setup/configureMonitorVM.sh
